requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
duckduckgo-search
google-generativeai
python-dotenv